Got rid of some warnings

+++ ios/Camp Joy Outdoors/CJOTreeInfoViewController.m
@@ -60,7 +60,7 @@
 }
 
 #pragma mark - UITableViewDataSource;
-- (int) numberOfSectionsInTableView:(UITableView *)tableView {
+- (NSInteger) numberOfSectionsInTableView:(UITableView *)tableView {
     return 3;
 }
 
@@ -75,7 +75,7 @@
     
 }
 
-- (int) tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section {
+- (NSInteger) tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section {
     if(section == 0) {
         return self.tree.tableData.count;
     } else {
@@ -83,7 +83,7 @@
     }
 }
 
-- (float) tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath {
+- (CGFloat) tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath {
     if(indexPath.section == 0) {
         return 48;
     } else {
